feat: move numpy to reproducibility optional dependencies
