+++ pyproject.toml
@@ -23,6 +23,9 @@ dependencies = [
 contrib = [
     "tqdm>=4.67.1",
 ]
+reproducibility = [
+    "numpy>=2.3.1",
+]
 
 [build-system]
 requires = ["hatchling"]
