+++ src/blazefl/core/__init__.pyi
@@ -1,6 +1,7 @@
 from blazefl.core.client_trainer import BaseClientTrainer as BaseClientTrainer, ProcessPoolClientTrainer as ProcessPoolClientTrainer, ThreadPoolClientTrainer as ThreadPoolClientTrainer
 from blazefl.core.model_selector import ModelSelector as ModelSelector
-from blazefl.core.partitioned_dataset import PartitionedDataset as PartitionedDataset
+from blazefl.core.partitioned_dataset import FilteredDataset as FilteredDataset, PartitionedDataset as PartitionedDataset
 from blazefl.core.server_handler import BaseServerHandler as BaseServerHandler
+from blazefl.core.utils import deserialize_model as deserialize_model, move_tensor_to_shared_memory as move_tensor_to_shared_memory, serialize_model as serialize_model
 
-__all__ = ['BaseClientTrainer', 'ProcessPoolClientTrainer', 'ThreadPoolClientTrainer', 'ModelSelector', 'PartitionedDataset', 'BaseServerHandler']
+__all__ = ['BaseClientTrainer', 'FilteredDataset', 'ProcessPoolClientTrainer', 'ThreadPoolClientTrainer', 'ModelSelector', 'PartitionedDataset', 'BaseServerHandler', 'serialize_model', 'deserialize_model', 'move_tensor_to_shared_memory']
