+++ src/blazefl/core/utils.pyi
@@ -1,4 +1,6 @@
 import torch
+from typing import Any
 
+def move_tensor_to_shared_memory(obj: Any, max_depth: int = 1) -> None: ...
 def serialize_model(model: torch.nn.Module, cpu: bool = True) -> torch.Tensor: ...
 def deserialize_model(model: torch.nn.Module, serialized_parameters: torch.Tensor) -> None: ...
