--- src/blazefl/utils/__init__.py
@@ -1,28 +1,0 @@
-"""
-Utilities module for BlazeFL.
-
-This module provides various utility classes and functions for the BlazeFL framework,
-including dataset manipulation, random seed control,
-and model serialization/deserialization.
-"""
-
-from blazefl.utils.dataset import FilteredDataset
-from blazefl.utils.ipc import move_tensor_to_shared_memory
-from blazefl.utils.seed import (
-    RandomStateSnapshot,
-    RNGSuite,
-    create_rng_suite,
-    seed_everything,
-)
-from blazefl.utils.serialize import deserialize_model, serialize_model
-
-__all__ = [
-    "serialize_model",
-    "deserialize_model",
-    "FilteredDataset",
-    "move_tensor_to_shared_memory",
-    "seed_everything",
-    "RandomStateSnapshot",
-    "create_rng_suite",
-    "RNGSuite",
-]
